Create Crime
There is a continuous discussion about whether or not the [illegal] immigrants infiltrating the USA from Mexico positively affect the crime rate in the border neighborhoods. Your assignment is to cast some light on the situation using data from Wikipedia.
Write a program that shall:
1. Download and extract crime-related data (namely, the total violent crime rate per 100,000 people) from the major US cities from the List of US cities by crime rate. https://en.wikipedia.org/wiki/List_of_United_States_cities_by_cr ime_rate
2. For each city, follow the link to its Wikipedia page and extract its coordinates. Calculate the distance to the border as the smallest distance from each city to San Ysidro, Yuma, Tucson, El Paso, Laredo, Del Rio, and Brownsville, TX. You are allowed to hardcode the coordinates of these locations. Use the great- circle distance formula.
3. Save the city names, crime rates, and smallest distances to the border in a 3-column CSV file.
4. Calculate and display the correlation between the two numbers (use scipy.corrcoeff()).
Finally, scatter plot crime rates against the distances. You can use Excel, Google Sheets, LibreOffice, or matplotlib. Add a trend line if you want. Write a one-page report that explains in layperson’s terms the procedure of your discovery, your findings (especially the correlation), and their significance. Include the scatter plot in the report. For the report, use MS Word, Google Docs, or LibreOffice.
Save each downloaded HTML document into the directory data. If the
directory does not exist, create it. If a needed document has been
previously downloaded, it shall be retrieved from the file, not from
1 the Web .